+++ src/main/java/com/penta/penta_service_posts/service/ReportService.java
@@ -1,12 +1,16 @@
 package com.penta.penta_service_posts.service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 import org.springframework.data.domain.Sort;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
+import com.penta.penta_service_posts.Converter.SpringSecurityAuditorAware;
 import com.penta.penta_service_posts.domain.Report;
+import com.penta.penta_service_posts.domain.Users;
 import com.penta.penta_service_posts.model.ReportDTO;
 import com.penta.penta_service_posts.repos.ReportRepository;
 import com.penta.penta_service_posts.util.NotFoundException;
@@ -16,9 +20,11 @@ import com.penta.penta_service_posts.util.NotFoundException;
 public class ReportService {
 
     private final ReportRepository reportRepository;
+    private final SpringSecurityAuditorAware auditorAware;
 
-    public ReportService(final ReportRepository reportRepository) {
+    public ReportService(final ReportRepository reportRepository , final SpringSecurityAuditorAware auditorAware ) {
         this.reportRepository = reportRepository;
+        this.auditorAware = auditorAware ;
     }
 
     public List<ReportDTO> findAll() {
@@ -42,13 +48,27 @@ public class ReportService {
 
     public void update(final UUID id, final ReportDTO reportDTO) {
         final Report report = reportRepository.findById(id)
-                .orElseThrow(NotFoundException::new);
-        mapToEntity(reportDTO, report);
-        reportRepository.save(report);
+            .orElseThrow(NotFoundException::new);
+        Optional<Users> currentUser = auditorAware.getCurrentAuditor();
+        if(currentUser.isPresent() && report.getCreatedBy().getId().equals(currentUser.get().getId())){
+            mapToEntity(reportDTO, report);
+            reportRepository.save(report);
+        }
+        else {
+            throw new AccessDeniedException("You are not allowed to update this Report.");
+        }
     }
 
     public void delete(final UUID id) {
-        reportRepository.deleteById(id);
+        final Report report = reportRepository.findById(id)
+            .orElseThrow(NotFoundException::new);
+                Optional<Users> currentUser = auditorAware.getCurrentAuditor();
+        if(currentUser.isPresent() && report.getCreatedBy().getId().equals(currentUser.get().getId())){
+            reportRepository.deleteById(id);
+        }
+        else {
+            throw new AccessDeniedException("You are not allowed to Delete this Report.");
+        }
     }
 
     private ReportDTO mapToDTO(final Report report, final ReportDTO reportDTO) {
@@ -66,7 +86,6 @@ public class ReportService {
         report.setPost(reportDTO.getPost());
         report.setReason(reportDTO.getReason());
         report.setReasonText(reportDTO.getReasonText());
-        report.setCreatedBy(reportDTO.getCreatedBy());
         return report;
     }
